# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "trip API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/trips": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Create Trip",
        "operationId": "create_trip_api_trips_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripSpecIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get Trip",
        "operationId": "get_trip_api_trips__trip_id__get",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "trips"
        ],
        "summary": "Patch Trip",
        "operationId": "patch_trip_api_trips__trip_id__patch",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripMetaPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "trips"
        ],
        "summary": "Delete Trip",
        "operationId": "delete_trip_api_trips__trip_id__delete",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}/days/{day_index}/items": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Add Item",
        "operationId": "add_item_api_trips__trip_id__days__day_index__items_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "day_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Day Index"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}/items/{item_id}": {
      "delete": {
        "tags": [
          "trips"
        ],
        "summary": "Remove Item",
        "operationId": "remove_item_api_trips__trip_id__items__item_id__delete",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "trips"
        ],
        "summary": "Patch Item",
        "operationId": "patch_item_api_trips__trip_id__items__item_id__patch",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}/days/{day_index}/order": {
      "put": {
        "tags": [
          "trips"
        ],
        "summary": "Reorder",
        "operationId": "reorder_api_trips__trip_id__days__day_index__order_put",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "day_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Day Index"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}/days/{day_index}/optimize": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Optimize Day",
        "description": "FR-8 — 순서 최적화. 캐시가 적중하면 외부 호출 없이 끝난다 (BR-29).",
        "operationId": "optimize_day_api_trips__trip_id__days__day_index__optimize_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "day_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Day Index"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}/items/{item_id}/opening-hours": {
      "put": {
        "tags": [
          "trips"
        ],
        "summary": "Set Opening Hours",
        "description": "BR-35 — 영업시간은 **여기서만** 채워진다. 외부 API·LLM 경로는 없다.",
        "operationId": "set_opening_hours_api_trips__trip_id__items__item_id__opening_hours_put",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpeningHoursIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BudgetLevel": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH"
        ],
        "title": "BudgetLevel"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "latitude": {
            "type": "number",
            "maximum": 39.0,
            "minimum": 33.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 132.0,
            "minimum": 124.0,
            "title": "Longitude"
          },
          "category_raw": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Raw"
          },
          "road_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Road Address"
          },
          "address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "stay_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 720.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Stay Minutes"
          },
          "memo": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Memo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "latitude",
          "longitude"
        ],
        "title": "ItemCreate",
        "description": "FR-5 — 검색 결과에서 항목 추가."
      },
      "ItemPatch": {
        "properties": {
          "stay_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 720.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Stay Minutes"
          },
          "memo": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Memo"
          },
          "travel_mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TravelMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "time_fixed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Fixed"
          },
          "fixed_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fixed Time"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ItemPatch",
        "description": "FR-7, FR-11 — 항목 편집."
      },
...
      "OpeningHoursIn": {
        "properties": {
          "weekday_rules": {
            "items": {
              "$ref": "#/components/schemas/DayRuleIn"
            },
            "type": "array",
            "maxItems": 7,
            "title": "Weekday Rules"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "weekday_rules"
        ],
        "title": "OpeningHoursIn"
      },
...
      "OptimizeIn": {
        "properties": {
          "anchor_start": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 36
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchor Start"
          },
          "anchor_end": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 36
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchor End"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "OptimizeIn",
        "description": "FR-8 — 순서 최적화 제약 (BR-19)."
      },
...
      "ReorderIn": {
        "properties": {
          "item_ids": {
            "items": {
              "type": "string",
              "maxLength": 36
            },
            "type": "array",
            "maxItems": 15,
            "title": "Item Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "item_ids"
        ],
        "title": "ReorderIn"
      },
...
      "TravelMode": {
        "type": "string",
        "enum": [
          "WALK",
          "CAR",
          "TRANSIT"
        ],
        "title": "TravelMode"
      },
      "TripDayOut": {
        "properties": {
          "day_index": {
            "type": "integer",
            "title": "Day Index"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItineraryItemOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "day_index",
          "items"
        ],
        "title": "TripDayOut"
      },
      "TripMetaPatch": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TripMetaPatch"
      },
      "TripOut": {
        "properties": {
          "trip_id": {
            "type": "string",
            "title": "Trip Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "destination": {
            "type": "string",
            "title": "Destination"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "party_size": {
            "type": "integer",
            "title": "Party Size"
          },
          "style_tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Style Tags",
            "default": []
          },
          "day_start_time": {
            "type": "string",
            "format": "time",
            "title": "Day Start Time"
          },
          "day_end_time": {
            "type": "string",
            "format": "time",
            "title": "Day End Time"
          },
          "default_travel_mode": {
            "$ref": "#/components/schemas/TravelMode"
          },
          "budget_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/TripDayOut"
            },
            "type": "array",
            "title": "Days"
          },
          "unresolved": {
            "items": {
              "$ref": "#/components/schemas/UnresolvedOut"
            },
            "type": "array",
            "title": "Unresolved",
            "default": []
          },
          "share_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Share Token"
          }
        },
        "type": "object",
        "required": [
          "trip_id",
          "title",
          "destination",
          "start_date",
          "end_date",
          "party_size",
          "day_start_time",
          "day_end_time",
          "default_travel_mode",
          "days"
        ],
        "title": "TripOut"
      },
      "TripSpecIn": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Title"
          },
          "destination": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Destination"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "party_size": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Party Size",
            "default": 2
          },
          "style_tags": {
            "items": {
              "type": "string",
              "maxLength": 20
            },
            "type": "array",
            "maxItems": 8,
            "title": "Style Tags",
            "default": []
          },
          "day_start_time": {
            "type": "string",
            "format": "time",
            "title": "Day Start Time",
            "default": "09:00:00"
          },
          "day_end_time": {
            "type": "string",
            "format": "time",
            "title": "Day End Time",
            "default": "21:00:00"
          },
          "default_travel_mode": {
            "$ref": "#/components/schemas/TravelMode",
            "default": "TRANSIT"
          },
          "budget_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetLevel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "destination",
          "start_date",
          "end_date"
        ],
        "title": "TripSpecIn",
        "description": "FR-1 — 여행 생성 입력."
      },
...
    }
  }
}